refactoring classes
