--- src/helpers/filters.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-class Filters(Enum):
-    ASCENDING = 'asc'
-    DESCENDING = 'desc'
