--- src/schemas/vowel_count_schema.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-from typing import List
-
-
-class VowelCountSchema(BaseModel):
-    words: List[str]
